LIBHYDRA-47. Added UMD colors and logo.
Removed Mirador stylesheet from the application CSS, since Mirador is now going to be included via iframe and hosted on the IIIF server. Using the Blacklight-recommended bootswatch stylesheet customization method. Expanded the size of the brand header and adjusted the indent so that the app name ("Archelon") appears in the header next to the UMD Libraries logo.

https://issues.umd.edu/browse/LIBHYDRA-47