LIBHYDRA-105 Initial drafts of download/disabled/expired pages.
Disabled/expired pages return an HTTP status of 410 - Gone

Uses a separate layout from application, in order to provide
flexibility for the navbar.

Added ".navbar-brand-um div.nolink" to custom.css so that the
application title in the navigation bar would appear the same,
even if it was not a hyperlink. Did this since the "Retrieve" controller
does not have a natural "home" URL at which to point an
application title link.

https://issues.umd.edu/browse/LIBHYDRA-105