LIBHYDRA-105 Removed unused actions and views
We do not allow the creation or editing of DownloadUrl objects directly,
so removed the "new", and "create" methods (the "edit" method was
previously removed).

Removed the views associated with the removed actions.

Removed JSON views, because they are not used.

Updated controller tests, using a Mock to replace network activity.

https://issues.umd.edu/browse/LIBHYDRA-105