LIBHYDRA-111 Enable Download URL routes to contain a period
By default, Rails does not allow dynamic segments in a URL to contain a
period (.) as it is typically used to specify the expected format of
the response.

The Download URL controller contains a route that may contain a URL,
which has periods in the hostname portion. The "generate_download_url"
route was updated to allow periods.

https://issues.umd.edu/browse/LIBHYDRA-111