LIBHYDRA-179. Fixed "Lint/IneffectiveAccessModifier" offenses
Removed "private" access modifier, because it actually has no
effect on class methods.

Did not use any of the suggestions from Rubocop (i.e., using
"private_class_method", or wrapping in a "class << self", as
that would change the way the "find_or_create_from_auth_hash"
method called them, which seemed less readable.

Replaced with a comment indicating that the methods are not intended to
be used outside the class.

https://issues.umd.edu/browse/LIBHYDRA-179