LIBHYDRA-179. Fixed "Style/AndOr" offenses
Added disable comment for these offenses because there are subtle
precedence differences between "&&" and "and" as used in these
statements, and didn't want to risk changing the logic.

See https://stackoverflow.com/a/1426835 for some of the issues.

Also had to add "Metrics/LineLength" because the disable comment
extended past the nominal line length.

https://issues.umd.edu/browse/LIBHYDRA-179