LIBHYDRA-184. Fixed user reference in Bookmark class
The UMD implementation of the Bookmark class had a "belongs_to" of
"cas_user", while in the Blacklight version, the "belongs_to" was
"user".

While apparently worked in Rails 5.0 (not sure why), in Rails 5.2, it
was causing bookmark creation failures, because the user associated
with the bookmark could not be found.

The failure manifested itself when creating a bookmark, in that the
transaction would always rollback.

Modified the "belongs_to" relationship in the UMD Bookmark class to
be "user", but use "CasUser" as the class.

https://issues.umd.edu/browse/LIBHYDRA-184