LIBITD-1486. Updated "minitest-reporters" gem
Removed version restriction on "minitest-reporters" gem so that it could
be updated to a version that uses the "MINITEST_REPORTER" environment
variable. The "MINITEST_REPORTER" environment variable is used by the
Jenkinsfile to cause test results to be formatted in the JUnit format
understood by Jenkins.

Removed the "minitest-ci" gem, because it does not appear to be used.

https://issues.umd.edu/browse/LIBITD-1486