LIBHYDRA-257. Added "/*.key" to .gitignore
Added "/*.key" (and changed "*.pem" to "/*.pem) so that SSL key and
PEM files in the root directory are not included in Git. This will
hopefully prevent key/PEM files used in the local development
environment from being accidentally checked in.

Modified the "*.pem" to "/*.pem" to more accurately reflect the
comment.

https://issues.umd.edu/browse/LIBHYDRA-257