LIBHYDRA-257. Updated import job completion percentage calculation
Modified job completion percentage calculation for imports. Assumes
that "updated" + "unchanged" indicates the total number of entries
processed so far, while "total" is the total number of entries to
process.

Note: This currently always returns 100% because the "total" count is the
total number of entries processed so far, not the total number of
entries to process (see LIBFCREPO-777).

https://issues.umd.edu/browse/LIBHYDRA-257