LIBHYDRA-257. Added checks to controller to validate actions
Added checks to edit and update to prevent changes to jobs that are in
the "import" stage.

Added checks to "import' action to prevent reimports and imports of
jobs that have failed validation.

These checks should not be triggerable via the GUI, but were added as
as a guard against any errant/malicious attempts to access the actions
via direct URLs.

https://issues.umd.edu/browse/LIBHYDRA-257