LIBHYDRA-260. Removed "status" field from ExportJob
Removed "status" field from ExportJob, which was largely unused,
replacing where needed with "plastron_status".

Updated ExportJob sample data creation to use "plastron_status" instead
of "status". Due to the way the ExportJob model is generated in the
"reset_with_sample_data" Rake task, changed the task prerequisites from
"create migrate" to "reset", as otherwise the ExportJob model for the
Rake task contains fields (such as "plastron_operation_id") that are
removed in the migrations. Using "reset" ensures that only the
columns in the "schema.rb" file are included in the model.

Renamed "plastron_operation_status" CSS matchers to "plastron_status" to
better reflect the actual usages. Updated the views to reflect the new
CSS.

https://issues.umd.edu/browse/LIBHYDRA-260