LIBHYDRA-267. Updated Validation Panel to show same status as index
Added a "status_text" local to the "_import_job_validation_panel.rb"
helper, and modified callers to populate the variable from the
"status_text" method on the controller. This enables the progress
percentage to be shown on the "show" page when a job is in progress,
and generally makes the GUI more consistent.

https://issues.umd.edu/browse/LIBHYDRA-267