LIBHYDRA-270. Refactored ImportJobResponse to include headers
Refactored ImportJobResponse to include headers. Replaced
"json_pretty_print" method with "headers_pretty_print" and
"body_pretty_print" methods to give extra flexibility to the
views about how to display the JSON in the GUI.

Added "last_response" method to ImportJob model to return an
ImportJobResponse object. This should hopefully better insulate the
model and controllers from changes in the import job message formats.

Added "headers_json" method to StompClient, to enable the headers to
be stored in the database in JSON format, instead of Ruby serialization.

Modified "_import_job_validation_panel.erb" to display headers and body
in the "JSON" subpanel.

Updated the tests

https://issues.umd.edu/browse/LIBHYDRA-270