LIBHYDRA-276. Added authentication to WebSocket connection creation
Modified the "app/channels/application_cable/connection.rb" to only
allow authenticated users to create a connection.

Since the "Connection" class does not have access to the session object,
this required adding a "cas_user" signed cookie in the
SessionsController when the user is logged in (and deleting it when
the session is destroyed, or the user logs out).

This approach follows the suggestion in
https://stackoverflow.com/a/43970848

https://issues.umd.edu/browse/LIBHYDRA-276