LIBHYDRA-276. Refactored session/cookie settings made by CAS
To better support testing, refactored the code in the SessionsController
that sets the session and cookies to use a "CasAuthentication" service
object.

This enables the test_helper to set the same sessions and cookie values
without having to duplicate code.

Added additional tests to verify that session and cookies are properly
created and destroyed.

https://issues.umd.edu/browse/LIBHYDRA-276