LIBHYDRA-276. Simplified code by simply reloading page on message
Instead of providing the updated HTML in the broadcast message, modified
the JavaScript handler to simply reload the page.

This an admin's export job page to automatically update to display
new export jobs from other users. In tests, the page reload did not
affect the scroll position of the browser on the page.

https://issues.umd.edu/browse/LIBHYDRA-276