LIBHYDRA-279. Added check_for_pending/import_job_status_check methods
With import jobs, there can be a race condition between the validation
check and the page display, where the
"validation successful"/"validation failed" message is sent by
ImportJobRelayJob before the page finishes loading and is subscribed
to the channel. This is problematic, because there is no subsequent
message to update the import job status, so the import job is never
updated in the GUI, and remains showing "Pending"/"In Progress".

Added the "check_for_pending" method to "import_jobs.coffee" to
determine if any displayed job is in a "pending"-type state, and if
so triggers the "import_job_status_check" method on ImportJobsChannel.
This "import_job_status_check" method checks the current status of the
jobs against the list sent by the client, and triggers a broadcast
message if there are differences. This fixes the race condition by
ensuring that a message update will be sent.

Also corrected naming of "channel_name"/"admins_channel_name" methods
to "stream"/"admins_stream", because these are streams, not channels
in Rails terminology.

Added test cases.

https://issues.umd.edu/browse/LIBHYDRA-279