LIBHYDRA-279. Modified import job updating for "show" and "edit" pages
Converted "import_jobs.coffee" to "import_jobs.js" because working in
CoffeeScript became more difficult that just working in plain JavaScript.

Modified the "received" function to conditionally reload the page, based
on whether or not the given data includes job ids present on the page.
If the "only-update-on-match" attribute is present on the page, the
page will only reload if job id is present in the page. Otherwise the
page will always reload (this ensures that new import jobs will be
shown).

Modified "ImportJobsChannel" to split jobs into per-user groups, so
that user-specific streams only get the updates visible to that user.

Modified "ImportJobRelayJob" to always send an array, as that is what
is expected by the "broadcast" method in "ImportJobsChannel".

Added additional documentation.

https://issues.umd.edu/browse/LIBHYDRA-279