LIBHYDRA-285. Rearranged order of actions in test
Rearranged the order of actions in the tests, as "create" action causes
a new Datatype to be created that mucks with the remaining actions.

Also fixed parameters sent to update to be "vocabulary_id" instead of
"vocabulary".

Placed "create" action at the end of the tests, so that it would not
interfere.

https://issues.umd.edu/browse/LIBHYDRA-285