LIBHYDRA-304. Added "status_update" route for STOMP listener
Added "status_update" routes to import and export job controllers which
are called by the STOMP listener to trigger an update message to the
import/export ActionCable channels.

This is needed because the web browser is connected to the Archelon
application, not the STOMP listener, so the "commit" hook in the
ImportJob and ExportJob models that triggers the
ImportJobRelayJob/ExportJobRelayJob has no effect. The "status_update"
routes on the import/export job controllers are called by the STOMP
listener to trigger the ImportJobRelayJob or ExportJobRelayJob, a
 appropriate, to send out the update notification to the ActionCable
channels.

Added "ARCHELON_URL" to "env_example", as the STOMP listener needs to
know the URL of the Archelon server to contact.

The "status_update" actions in the controllers skip authentication to
simplify the STOMP listener implementation, and because it seems
unlikely that they can be abused.

Modified "app/assets/javascripts/channels/import_jobs.js" to request an
update when any job is in an "in progress" state (instead of only for
jobs that were "in progress" and in the "validate" stage), because job
updates may come in quick enough succession that a notification may be
missed when the page is reloading.

https://issues.umd.edu/browse/LIBHYDRA-304