LIBHYDRA-307. Added Rake tasks for adding public keys for users
Created two Rake tasks in "lib/tasks/add_public_key.rake" - one to take
the public key as a string, the other to read the public key from a
file:

* rails user:add_public_key[cas_directory_id,public_key]
* rails user:add_public_key_file[cas_directory_id,public_key_file]

The tasks assume:

* that the user exists
* the the public key does not exist
* that the public key file is readable

A message will be printed and the task will return with an exit status
of 1 if any of the assumptions are not true.

If the public key is added, a message will be printed, and the exit
status will be 0.

Updated the README.md with a "Rake Tasks" section.

https://issues.umd.edu/browse/LIBHYDRA-307