LIBHYDRA-311. Tweaked MIME type translations
Modified MIME type translations in "en.yml" to include the actual
MIME type in parentheses.

Did this here, instead of in the code, because there are multiple "XML"
MIME types, and if both had an "XML" translation (which would be
disambiguated by the parenthetical MIME type added in the code), there
is the risk that we wouldn't do that somewhere and so end up with
an ambiguous MIME type.

https://issues.umd.edu/browse/LIBHYDRA-311