LIBHYDRA-314. Added "remove" functionality to TextField
Added "handleRemove" method to implement removal of "LiteralValue"
objects. This included adding a "keyCounter" class variable, that
increments whenever a "LiteralValue" is added to provide as a
React "key" field, so that rendering works properly when an
object is removed.

Removed "bind" of "handleAdd" method, replacing with the "arrow" syntax,
so that it would be consistent with the "handleRemove" method.

Modified the "Add" button be be hidden and disabled when items can no
longer be added (i.e., maximum item count is reached).

https://issues.umd.edu/browse/LIBHYDRA-314