LIBHYDRA-315. Added PlainLiteral, TypedLiteral, and Repeatable
Replaced the "LiteralValue" and "TextField" React components with a
"PlainLiteral" React component.

Added a "TypedLiteral" React component for associating a datatype
with a textbox.

Added "Repeatable" component to enable React components to be
"repeatable", i.e., enable multiple instances to be created/removed
dynamically.

Added "RepeatablePlainLiteral" and "RepeatableTypedLiteral" components
because of a "react-rails" limitation that doesn't allow anonymous
JavaScript functions to be passed to React.

Added an "Introduction" section to the React Styleguidist documentation,
as well as examples of each component.

https://issues.umd.edu/browse/LIBHYDRA-315