LIBHYDRA-318. Modified TypedLiteral to generate proper SPARQL syntax
Before this change, the TypedLiteral was generating the following
SPARQL query when an "edtf" date was changed:

```
DELETE {
  <SUBJECT_URI> <http://purl.org/dc/elements/1.1/date> "1948-08-11"@http://id.loc.gov/datatypes/edtf .
}
INSERT {
  <SUBJECT_URL> <http://purl.org/dc/elements/1.1/date> "1948-08-12"@http://id.loc.gov/datatypes/edtf .
} WHERE {}
```

(where SUBJECT_URI is the URI is the item being modified)

This is apparently not the correct SPARQL format for typed object, which
expects:

```
DELETE {
  <SUBJECT_URI> <http://purl.org/dc/elements/1.1/date> "1948-08-11"^^<http://id.loc.gov/datatypes/edtf> .
}
INSERT {
  <SUBJECT_URL> <http://purl.org/dc/elements/1.1/date> "1948-08-12"^^<http://id.loc.gov/datatypes/edtf> .
} WHERE {}
```

Note the change in the "edtf" type syntax.

Could not find a way to generate this syntax from the DataFactory,
so just put the value together manually.

https://issues.umd.edu/browse/LIBHYDRA-318