LIBHYDRA-318. Initial implementation of update validation
Updated ResourceController to send the "PlastronArg-validate" and
"PlastronArg-model" headers to Plastron, so that validation can
be performed on the update request.

Added the "plastron_model_from_rdf_type" method to determine the model
to send in the header, based on the items.

Modified the "update" method to retrieve the resource from Fedora
(in order to determine the model), and handle any validation errors
that come back from Plastron.

Added additional JavaScript code and view helpers to display any
validation errors. This code is currently extremely simplistic, and
it is expected that better error handling (such as field highlighting)
will be added in the future.

https://issues.umd.edu/browse/LIBHYDRA-318