LIBHYDRA-363. Fix timeout error handling
Fixed display of timeout errors. Prior to this change, the error was
not displayed, because of an undefined "errors" variable in the
"resource.js" JavaScript. Corrected the timeout error handling to
ensure that a message (defined in the en.yml file) is displayed.

In "resource_controller.rb" modified the generation of the timeout
error to be passed back as a JSON string, to match the errors coming
from Plastron.

https://issues.umd.edu/browse/LIBHYDRA-363