LIBHYDRA-366. Override Blacklight helper methods to display Fedora title
Created utility methods "display_title" and "sort_titles_by_language" in
"app/services/resource_service.rb" to generate a display title from a
resource (the title is assumed to use the
"http://purl.org/dc/terms/title" predicate). The titles are sorted by
language, to ensure that a consistent order when there are multipl
 titles.

Used "app/helpers/blacklight_helper.rb" to override two Blacklight
helper methods to use the ResourceServices utility methods to generate
the title. If the title cannot be generated, the default Blacklight
methods (which use Solr) are used.

This causes the resource detail page to immediately display the correct
title after it has been editing, instead of waiting for Solr to process
the change.

https://issues.umd.edu/browse/LIBHYDRA-366