LIBHYDRA-410. Added ImportJob methods for structure and relpath
Added methods to ImportJob for returning the collection structure
(either ":flat" or ":hierarchical"), as well as for returning the
"relpath" of the collection.

The "relpath", despite being a "relative path" will always have an
initial "/", as that seems to be what is expected by Plastron.

Added a "with_constant" method to "test/test_helper.rb" to allow the
"FCREPO_BASE_URL" constant defined in the config/initializers/fcrepo.rb"
to be easily redefined for tests.

https://issues.umd.edu/browse/LIBHYDRA-410