LIBFCREPO-977. Updated Archelon Docker config for SFTP private key
Added an SFTP server to the Archelon Docker image, and created a
"bin/docker_start.sh" script to start both the SFTP server and
Archelon.

Modified the "config/export.yml" configuration to use EXPORT_DIR and
EXPORT_BASE_DESTINATION environment variables, if present.

Updated the Archelon Docker config to enable a public key to be
provided in the "PLASTRON_PUBLIC_KEY" to allow SFTP communication with
Plastron.

Updated the "docs/ArchelonDevelopmentEnvironmentDocker.md" instructions.

https://issues.umd.edu/browse/LIBFCREPO-977