LIBFCREPO-977. Added "REPO_EXTERNAL_URL" configuration param to Archelon
Added optional "REPO_EXTERNAL_URL" configuration parameter to Archelon

This is needed so that the relative path (relpath) of a resource can be
determined for imports.

The "collection_relpath" method in "app/models/import_job.rb" was
modified to use REPO_EXTERNAL_URL (if present) to return the correct
relpath.

Added "REPO_EXTERNAL_URL" to the "docker-archelon-template.env" template
configuration file.

Modified "config/imports.yml" to make use of the "IMPORT_BINARIES_DIR"
and "IMPORT_BINARIES_BASE_LOCATION" in the "default" configuration
(which is used by the "development" environment") if those parameters
are present in the environment. This enables them to be defined when
using Docker.

https://issues.umd.edu/browse/LIBFCREPO-977