LIBHYDRA-429. Bumped Rails version to 5.2.6
Updated Rails version to 5.2.6, to fix the "mimemagic" issue that
was preventing the Docker image from being built.

Ran

```
> bundle update rails --patch
```

to update the gems. The "–patch" flag caused Bundler to prefer upgrading
dependencies to the next patch version (where possible), instead of
automatically going to the latest version of the dependency (which could
involve a major version change).

https://issues.umd.edu/browse/LIBHYDRA-429