LIBHYDRA-429. Updated puma from v4.3.5 to v4.3.8
Updated "puma" gem from v4.3.5 to v4.3.8 to fix a "bundler-audit"
warning.

Ran

```
> bundle update puma --patch
```

to update the gem. The "–patch" flag caused Bundler to prefer upgrading
dependencies to the next patch version (where possible), instead of
automatically going to the latest version of the dependency (which could
involve a major version change).

https://issues.umd.edu/browse/LIBHYDRA-429