LIBHYDRA-429. Updated rdoc from v6.2.1 to v6.3.1
Updated "rdoc" gem from v6.2.1 to v6.3.1 to fix a "bundler-audit"
warning.

Ran

```
> bundle update rdoc --minor
```

to update the gem. The "–minor" flag caused Bundler to prefer upgrading
dependencies to the next minor version (where possible), instead of
automatically going to the latest version of the dependency (which could
involve a major version change).

https://issues.umd.edu/browse/LIBHYDRA-429