LIBHYDRA-428. Updated logging to respect RAILS_LOG_TO_STDOUT env var
In LIBHYDRA-215, the Puma configuration was updated to send
STDOUT/STDERR to "log/puma.log" and "log/puma_error.log".

This interferes with the Kubernetes practice of dumping logs to
STDOUT, so added a check to disable the logging to the files if the
"RAILS_LOG_TO_STDOUT" environment variable is set.

https://issues.umd.edu/browse/LIBHYDRA-428