LIBHYDRA-377. Updated ControlledURIRef to display non-vocabulary terms
Modified the ControlledURIRef React component to display a URI that is
not in the vocabulary simply as the URI, instead of showing a blank
value.

The non-vocabulary term is added to the drop-down list, as so will be
available to the user, at least until the field is submitted with a
different value.

https://issues.umd.edu/browse/LIBHYDRA-377