LIBHYDRA-443. Added PlastronService, JsonRestResult, and ImportJobInfo
Added PlastronService to retrieve information from the Plastron REST
endpoint as a JsonRestResult object.

The JsonRestResult object is intended to encapsulate both success
and failure of the server request (success returning a Hash of the
parsed JSON, failure returning an error message).

ImportJobInfo is a customization of the JsonRestResult object to be
more in line with what is needed for the view.

Updated "show" method in ImportJobsController to query the
PlastronService.

https://issues.umd.edu/browse/LIBHYDRA-443