LIBHYDRA-449. Update export job status dynamically without page reload
Largely followed the import status example in LIBHYDRA-446.

One issue was that the "status" field in the "_export_job_status.erb"
helper actually consists of two "td" elements – one for the status,
the other for the "Download" button. This made it problematic to update
the status because targeting the "td" element for the "status" field for
updates resulted in nesting both "td" elements from the helper in a
single "td" element in the table (because the HTML attribute being used
to locate the element in JavaScript was inside the first "td" element).

Ideally, it would have been nice to be able to group the two "td"
elements within a single HTML element that had an attribute that could
be targeted by JavaScript so that they could be replaced together, but
I could not find a way to do this. Instead, added a
"_export_job_table_row.erb" helper that replaces the entire row in the
table when the status is updated.

https://issues.umd.edu/browse/LIBHYDRA-449