LIBHYDRA-467. Fixed "assets:precompile" database adapter in Dockerfile
The Archelon Dockerfile had the following line to run the
"assets:precompile" task:

```
RUN PROD_DATABASE_ADAPTER=postgresql bundle exec rails assets:precompile
```

The "PROD_DATABASE_ADAPTER" environment variable was believed to be
needed for the "assets:precompile" task to successfully run, but there
is no other reference to the "PROD_DATABASE_ADAPTER" variable anywhere
else in the codebase. This was likely copy-and-pasted from some
other Rails application (such as "student-applications").

The environment variable has been in the codebase for quite some time,
but was apparently unused. When the "delayed_job" gem was added in
LIBHYDRA-462, the "assets:precompile" task started causing the following
error when building the Docker image:

```
 > [8/8] RUN bundle exec rails assets:precompile:
...
```

Changed the "PROD_DATA_ADAPTER" environment variable to
"ARCHELON_DATABASE_ADAPTER" to match the environment variable expected
by the "config/database.yml" file, which corrected the error.

https://issues.umd.edu/browse/LIBHYDRA-467