LIBHYDRA-465. Moved Plastron-specific code out of StompService
Moved the Plastron-specific code out of Plastron and into the
ResourceController class.

The StompService is actually communicating with Plastron through
ActiveMQ, and it is capable to sending message to any queue
(not just those associated with Plastron).

Since the ResourceController is the class that knows about Plastron,
it seemed to be better to place the Plastron-specific information
in that class.

https://issues.umd.edu/browse/LIBHYDRA-465