LIBHYDRA-564. Handle "one term" vocabularies
When a vocabulary only has one term, there is no "@graph" element in
the JSON returned from the vocabulary server. Once a second term is
added, the "@graph" element is added, and populated with the
terms.

For example, a "collection" vocabulary with only one term looks like:

```
{
  "@context": {
    "owl": "http://www.w3.org/2002/07/owl#",
    "rdfs": "http://www.w3.org/2000/01/rdf-schema#",
    "dc": "http://purl.org/dc/elements/1.1/"
  },
  "@id": "http://vocab.lib.umd.edu/collection#0001-GDOC",
  "dc:identifier": "0001-GDOC",
  "rdfs:label": "United States Government Posters"
}
```

while a "collection" vocabulary with two terms looks like:

```
{
  "@context": {
    "owl": "http://www.w3.org/2002/07/owl#",
    "rdfs": "http://www.w3.org/2000/01/rdf-schema#",
    "dc": "http://purl.org/dc/elements/1.1/"
  },
  "@graph": [
    {
      "@id": "http://vocab.lib.umd.edu/collection#0211-UA",
      "rdfs:label": "Diamondback Photos",
      "dc:identifier": "0211-UA",
      "owl:sameAs": {
        "@id": "http://vocab.lib.umd.edu/collection#0211-UA"
      }
    },
    {
      "@id": "http://vocab.lib.umd.edu/collection#0001-GDOC",
      "rdfs:label": "United States Government Posters",
      "dc:identifier": "0001-GDOC"
    }
  ]
}
```

While we don't expect to have "one term" vocabularies very often, they
may occur when a new vocabulary is being rolled out, so it was decided
to handle that case.

These changes modify the "parse_options" method in
"app/services/vocab_service.rb" to handle both "one term" and multi-term
vocabularies.

Added a test to verify the "one term" vocabulary.

https://umd-dit.atlassian.net/browse/LIBHYDRA-564