LIBHYDRA-572. Update ControlledURIRef to handle unsetting value
Modified the ControlledURIRef React component to disable the hidden
"insert" field when the user selects the "empty" option in drop-down
list.

This is done so that when the "empty" option is chosen, the
generated SPAQRL only contains a "DELETE" statement.

Prior to this change the SPARQL query should generate both a "DELETE"
statement, and an "INSERT" statement of the form:

```
INSERT { <subjectURI> <predicateURI> <> . }
```

which resulted in the <subjectURI> being used for the node field.

This change fixes the validation issue seen in Plastron, because no
value is being inserted, so validator that checks whether the new value
is in the vocabulary is not triggered.

https://umd-dit.atlassian.net/browse/LIBHYDRA-572