LIBHYDRA-572. Remove duplicate "empty" option on ControlledURIRef fields
Modified the ControlledURIRef component to remove a duplicate "empty"
option that occurs when the initial value of field is the empty value.

This was occurring because the "empty" value is not in the vocabulary,
and so gets added to the vocabulary in the constructor, while the
React component template also contains an "empty" option field.

Modified the ControlledURIRef component so that the initial field value
does not get added to the vocabulary if it is an empty string.

https://umd-dit.atlassian.net/browse/LIBHYDRA-572