LIBHYDRA-573. Added "Vocab" object to replace "Vocabulary" model
Added a "Vocab" object that retrieves vocabulary information from
the vocabulary server.

Methods have been added to the "Vocab" object to make it a "drop-in"
replacement of the Vocabulary model in the non-Vocabulary code.

Refactored "VocabService" to support the "Vocab" object, pulling out the
"vocab_options_hash_by_identifier" method to retrieve an unfiltered
vocabulary based solely on the vocabulary identifier.

https://umd-dit.atlassian.net/browse/LIBHYDRA-573