LIBHYDRA-573. Deleted "public/published_vocabularies" directory
The "public/published_vocabularies" directory is no longer needed,
and so has been removed.

Also removed the "public/published_vocabularies" configuration in
the ".gitignore" file, to highlight any files the user might have for
deletion.

Left the "public/published_vocabularies" configuration in the
".dockerignore" file for the moment, to ensure that any lingering
contents in that directory will never be added to a Docker image.

https://umd-dit.atlassian.net/browse/LIBHYDRA-573