LIBHYDRA-77. Limit initial number of pcdm:members displayed
Some items have tens or hundreds of "pcdm:members" (such as
<https://archelon.lib.umd.edu/catalog/https:%2F%2Ffcrepo.lib.umd.edu%2Ffcrepo%2Frest%2Fdc%2F2023%2F1%2F31%2F26%2F3f%2F35%2F31263f35-6122-47d7-8c95-5566e5cbdee5>)

Since appears to be a "one-off" display issue, implemented the display
handling within the "_show_default.html.erb" file itself, instead of
doing anything fancier. The initial display limit of "8" is also
hard-coded for the same reason.

In the "_show_default.html.erb" file, split the "pcdm_members" subset
array into an "initial" list of (up to) 8 items, and a "more" list for
any additional items. The "initial" list is display normally as "<li>"
items in a "<ul>" unordered list.

If the "more" list has items, an HTML "Details disclosure" element is
used with a "summary" of "More" to display the remaining items.

https://umd-dit.atlassian.net/browse/LIBHYDRA-77