LIBHYDRA-571. Enable ResouceHelper.get_field_values to distinguish datatypes
The "get_field_values" method in "resource_helper.rb" was not properly
distinguishing item values based on the predicate and datatype.
Instead, it would return all the values for a given predicate,
irrespective of its datatype (meaning that if two fields had the
same predicate, but different datatypes, each field would get both
values).

In fixing the "get_field_values" method to properly distinguish between
datatypes, the "uri_to_fieldname" method turned out to be unnecessary,
and the method signature itself was changed to only require an "item"
parameter, and a "field" parameter (so both the field's predicate and
datatype (if present) could be queried).

Updated the places where the "get_field_values" method was used with
the new method signature.

Added a unit test to verify that the "get_field_values" method now
properly distinguishes between datatypes.

https://umd-dit.atlassian.net/browse/LIBHYDRA-571