LIBHYDRA-571. Added specialized display formatting for "handle" field
Added "display_handle" method to "application_helper.rb" to control
the formatting of the "handle" field. Updated the "display_node" method
to delegate to the "display_handle" method when a field's datatype is
'http://vocab.lib.umd.edu/datatype#handle'

Added the "HANDLE_HTTP_PROXY_BASE" environment variable to the
"env_example" file, with a default of "http://hdl-local/" in the
".env.development" file. The "http://hdl-local/" server does not
actually exist -- it is merely intended as a placeholder.

Updated the unit tests, including refactoring to put more of the
setup into the "setup" method.

https://umd-dit.atlassian.net/browse/LIBHYDRA-571