LIBFCREPO-1439. Filter out vocab term exactly matching vocab URL
Moved the parsing of vocabulary "graph" objects into the "parse_graph"
method to enable the filtering out any "term" that has a "uri" that
exactly matches the URL of the vocabulary being downloaded. This removes
the extraneous the self-identifying vocabulary term.

Removed the filtering of the self-identifying vocabulary term from
the "vocab_options_hash" method, as it is no longer needed.

Specifically decided to pass the vocabulary identifier to the
"parse_graph" method, and using it to generate the vocabulary URI
to be filtered out (via the "generate_url" method), in the expectation
that at some point there may be a difference between the vocabulary
URI and the URL from which the vocabulary is retrieved.

https://umd-dit.atlassian.net/browse/LIBFCREPO-1439