LIBFCREPO-1440. Removed "Access Level" field from metadata edit page
Removed the "access" field from the Item, Poster, Letter, and Issue
content models. This remotes the "Access Level" field from the item
metadata edit page.

Updated the "test/services/vocabulary_service_test.rb", which was using
the "access" field from the content model in one of the tests to use
a locally-generated example of the field.

https://umd-dit.atlassian.net/browse/LIBFCREPO-1440