LIBFCREPO-1458. Added optional "edit_only" content model field
Added an optional "edit_only" field to the content model definition to
enable specific fields, such as "Presentation Set" to not be displayed
on the item detail page.

This is needed because otherwise the "Presentation Set" field would be
shown twice on the item detail page -- once for the Solr-based field
specified in the “app/controllers/catalog_controller.rb” configuration
and once for the Fedora-based field specified in the
“config/content_models.yml” file.

Based on June 26, 2024 WebEx discussions in the “Fedora” channel with
Josh Westgard, the preference is to display the Solr field on the
item detail page because:

> ... that way we know that the update has "stuck" all the way through
> the stack. So even though there is a delay associated with making
> such a change, in the long run this is better for display in Archelon.

Modified the "_show_default.html.erb" file to skip display of any fields
where the "edit_only" field is true.

Updated the "ContentModelDefinitions.md" documentation.

https://umd-dit.atlassian.net/browse/LIBFCREPO-1458