LIBFCREPO-1460. Fix "ControlledURIRef" delete when "Repeatable"
In LIBHYDRA-356, a "componentWillUnmount" method was added to the React
components to support deletion when used with the "Repeatable" React
component. The "ControlledURIRef" component apparently missed in the
update.

Added the "componentWillUnmount" to fix the issue with deleting entries
when the "ControlledURIRef" component is used with a "Repeatable" React
component.

https://umd-dit.atlassian.net/browse/LIBFCREPO-1460