LIBFCREPO-1467. Update to Rails v5.2.8.1
Update to Rails v5.2.8.1 (from Rail v5.2.6.2).

This was a patch version upgrade, and so was fairly straightforward,
except for the Rails v5.2.8.1 fix for CVE-2022-32224, which caused some
tests to fail, and the application home page to not display. This
required adding a workaround to the "config/application.rb" file
to configure the "Pysche" YAML processor to allow the
"ActiveSupport::HashWithIndifferentAccess" class during deserialization.

https://umd-dit.atlassian.net/browse/LIBFCREPO-1467