LIBFCREPO-1471. Fixed "UMD TODO" in catalog_controller for detail page
Fixed the "UMD TODO" markers in the "catalog_controller.rb" page that
was keeping the item detail page from displaying with all fields.

This necessitated adding helper methods to "application_helper.rb" from
Archelon.

The SolrDocument class in Blacklight 8 no longer supports direct
assignment, as it did in Blacklight 6, see
<https://github.com/projectblacklight/blacklight/commit/dcfc74aebd4446aac85b2c2f7d10cd2c6ff8fef3>

Instead, it should be converted into a Hash, which is done in
"from_subquery" method.

https://umd-dit.atlassian.net/browse/LIBFCREPO-1471