LIBFCREPO-1471. Added BlacklightHelper and set application name
The "BlacklightHelper" class in legacy Archelon was used to
set the application name, and provide other methods.

Blacklight expects the application name to be set in an I18n file,
so that was done here in "config/locales/blacklight.en.yml.

It is not clear whether any of the other methods in the
"BlacklightHelper" class are actually needed, so commenting them
all out for now. If the methods all remain commented out, the
"app/helpers/blacklight_helper.rb" file can be deleted.

https://umd-dit.atlassian.net/browse/LIBFCREPO-1471