LIBFCREPO-1471. Renamed "SendJSONRequest" to "SendJsonRequest"
Rails 7 uses the "zeitwerk" autoloader which is stricter about matching
class names to filenames. The "JSON" in "SendJSONRequest" was changed
to "Json" (i.e., "SendJsonRequest") to enable "zeitwerk" to find the
class.

See <https://guides.rubyonrails.org/v7.0/classic_to_zeitwerk_howto.html#does-my-application-comply-with-zeitwerk-conventions-questionmark>
for more information.

It was decided to rename this class (instead of using an “inflector”)
for simplicity (and for consistency with the “JsonRestResult” class.

Updated the reference to the "SendJsonRequest" class in
"vocabulary_service.rb".

https://umd-dit.atlassian.net/browse/LIBFCREPO-1471