LIBFCREPO-1471. Added "resource.js" script and updated form for AJAX
Added the "resource.js" script from legacy Archelon, and added to the
"application.js" file, so it would be loaded.

Changed the form in "edit.html.erb" so that the "data-remote: true"
attribute would be included in the HTML (via the "local: false"
parameter), so that the form would handle AJAX properly (this is needed
because in Rails 7, forms are no longer "remote" by default).

https://umd-dit.atlassian.net/browse/LIBFCREPO-1471