LIBFCREPO-1471. Added Download URL links to sidebar component template
Modified the Blacklight 8 sidebar component template to include
links for generating the Download URLs:

* “Select Page for More Options” on item detail pages with
  “pcdm_members”
* “Select File for Download Options” on item detail pages with
  “pcdm_files”
* “Download“/”Generate Download URL” links on item detail pages with an
   “rdf_type” containing “fedora:Binary”

These additions are largely the same as those in
“app/views/catalog/_show_sidebar.html.erb” from legacy Archelon, with
changes to the way field values are retrieved. For example, for
“pcdm_members”, instead of

```
pages = doc_presenter.field_value('pcdm_members')
```

the following needed to be used:

```
<% pcdm_member_field_value = presenter.field_value(presenter.configuration.show_fields['pcdm_members']) %>
<% pages = pcdm_member_field_value
Updated "routes.rb" with the Download URL routes from legacy Archelon.
```

https://umd-dit.atlassian.net/browse/LIBFCREPO-1471