LIBFCREPO-1471. Copied PlastronServices classes used for publication
Copied from legacy Archelon the PlastronServices classes used for
item publication.

Because of the Rails 7 "zeitwerk" class loader naming conventions,
renamed the "HTTP" module to "Http", so that the classes can be
found by the loader.

Also, in "send_activity_stream_message.rb", changed "SendJSONRequest" to
"SendJsonRequest", to enable the class to be found.

https://umd-dit.atlassian.net/browse/LIBFCREPO-1471