LIBFCREPO-1471. Added "rspec_stub_const" to fix skipped tests
In legacy Archelon, the “stub_const” method being used was from the
“RSpec::Mocks::ExampleMethods” module in the “rspec-mocks” gem.

In Rails 7, the "ActiveSupport::Testing::ConstantStubbing" module also
has a “stub_const” method, that appears to be getting called instead.

This was causing errors similar to the following, because the wrong
"stub_const" method was being called:

```

ERROR ImportJobsControllerTest#test_should_create_import_job (0.50s)
Minitest::UnexpectedError:         ArgumentError: wrong number of arguments (given 2, expected 3)
            test/test_helper.rb:132:in `mock_stomp_connection'
            test/controllers/import_jobs_controller_test.rb:69:in `block in <class:ImportJobsControllerTest>'
```

In order to make it clearer which “stub_const” method was being used,
added a “rspec_stub_const” method to the “test/test_helper.rb” file, and
updated the usages in the tests.

https://umd-dit.atlassian.net/browse/LIBFCREPO-1471