LIBFCREPO-1523. Import Jobs dynamic status update
Implemented dynamic status updates for Import Jobs.

Required minor modifications to use the "consumer" object in
"import_jobs_channel.js".

In "_import_job_status.erb" needed to add "local: false" to the
"form_with" stanza to deal with Rails 7.1 and CSRF tokens (this is done
so that a "data-remote: true" directive is used).

Note: In the local development environment, status updates do NOT work
in Firefox (possibly the "archelon-local" hostname is confusing it).
The functionality does work in Chrome and Safari.

https://umd-dit.atlassian.net/browse/LIBFCREPO-1523