LIBFCREPO-1550. Updated PingController with more reliable database check
Modified the database check in the PingController to more reliably
detect database connection issues. It is unclear why the check from
Archelon 1.16.0 (using Rails 5.2) is no longer reliable with Rails 7.1,
but in Kubernetes the "fcrepo-archelon" pod was consistently restarting
due to the PingController indicating that it could not connect to the
database.

https://umd-dit.atlassian.net/browse/LIBFCREPO-1550