LIBFCREPO-1551. Implemented OmniAuth "developer" strategy to bypass CAS
Using CAS authentication for Archelon in the local development
environment is problematic because DIT does not allow the “localhost”
hostname to communicate with the CAS server.

To work around this, Archelon uses the “archelon-local” hostname in the
local development environment.

This works for most Rails functionality, but appears to break Rails
ActionCable functionality in Firefox, where the web browser can never
subscribe to the ActionCable channels provided by the server.

This implementation uses the OmniAuth “developer” strategy (see
<https://github.com/omniauth/omniauth/blob/v1.9.2/lib/omniauth/strategies/developer.rb>)
which instead of going to CAS, presents a form with a single “Uid” field.

Note: Because Archelon uses the provided username to pull information
from LDAP, only “active” LDAP users can be used (for example “blparker”
for “Bria Parker” will not work, because she is presumably no longer in
LDAP, while “lshiota” for Lisa Shiota does work).

The OmniAuth “developer” strategy is only available in the Rails
“development” environment (`ENV["RAILS_ENV"] == "development"` and an
`ARCHELON_AUTH` environment variable set to `developer`. This should
make it impossible to accidentally trigger the authorization method in
production.

In order to use the "localhost" hostname, the “config.hosts” environment
variable in “config/environments/development.rb” must be changed to:

```
config.hosts = [
    "archelon-local",
    "localhost"
  ]
```

While possible to include both hostname all the time, it seemed
like that might prove confusing.

https://umd-dit.atlassian.net/browse/LIBFCREPO-1551